fix(s1.1): apply FIX-6..9 — parse_level_from, ensure_log_dir, verbose precedence, derive Default
