+++ crates/scmux-daemon/src/scheduler.rs
@@ -167,7 +167,7 @@ pub async fn poll_cycle(state: &Arc<AppState>) -> anyhow::Result<()> {
 }
 
 /// Returns true if the cron expression should fire within the current 15s window.
-pub(crate) fn should_run_now(expr: &str, now: &chrono::DateTime<Utc>) -> bool {
+pub fn should_run_now(expr: &str, now: &chrono::DateTime<Utc>) -> bool {
     let Ok(schedule) = Schedule::from_str(expr) else {
         return false;
     };
@@ -177,34 +177,4 @@ pub(crate) fn should_run_now(expr: &str, now: &chrono::DateTime<Utc>) -> bool {
         .next()
         .map(|t| t <= *now)
         .unwrap_or(false)
-}
-
-#[cfg(test)]
-mod tests {
-    use super::should_run_now;
-    use chrono::{TimeZone, Utc};
-
-    #[test]
-    fn td_05_should_run_now_true_when_cron_fires_in_window() {
-        let now = Utc
-            .with_ymd_and_hms(2026, 1, 1, 12, 0, 10)
-            .single()
-            .expect("valid datetime");
-        assert!(should_run_now("0 0 12 1 1 *", &now));
-    }
-
-    #[test]
-    fn td_06_should_run_now_false_when_cron_does_not_fire_in_window() {
-        let now = Utc
-            .with_ymd_and_hms(2026, 1, 1, 12, 0, 10)
-            .single()
-            .expect("valid datetime");
-        assert!(!should_run_now("0 1 12 1 1 *", &now));
-    }
-
-    #[test]
-    fn td_07_should_run_now_invalid_cron_returns_false() {
-        let now = Utc::now();
-        assert!(!should_run_now("not-a-cron", &now));
-    }
 }
